parse disambiguated pawn captures as san_move
Previously a fully-disambiguated pawn capture was confusingly parsed as
a lan_move node, meaning that clients of the library had to account for
both san_move and lan_move nodes to be robust.

After this change, a fully-disambiguated pawn capture is a san_move
node.  Clients do not need to worry about lan_move nodes unless there
is an actual desire to support LAN in the movetext.  (When supporting
LAN a mix of node types must always be accounted for, as there are a
few ambiguities which are resolved in favor of san_move.)