allow "N" Novelty glyphs without leading space
The "N" annotation glyph for Novelty is tough for the parser since "N"
also represents Knight in SAN.

Here we lift the requirement for space before "N" glyphs by embedding
just a little bit of knowledge of the rules of chess in the parser,
making a distinction between promoting and non-promoting ranks.  Since
promotion is mandatory when moving a pawn to a promoting rank, we can
know that "40. g8N" is an irregularly-notated underpromotion, not a
Novelty annotation.  "g8NN" is also parseable as a promotion followed
by an annotation.

The above assumes that "g8" alone in SAN does not imply a Queen
promotion.  That is: "g8N" cannot also mean "g8=QN", or the parser is
stuck with a true ambiguity.

For historical reasons, implied Queen promotion is still permitted on
LAN notation!  LAN support remains best-effort.

Space _after_ N annotations glyphs is still required, due to ambiguities
which are not resolvable without move generation.